Fix for validation errors in ME and ND (Recidiviz/recidiviz-data#29069)
## Description of the change

Fix for validation errors in ME and ND. This should fix:
-
https://app.zenhub.com/workspaces/all-pods-6324d1d66be1c200119ab65a/issues/gh/recidiviz/recidiviz-data/29038
-
https://app.zenhub.com/workspaces/all-pods-6324d1d66be1c200119ab65a/issues/gh/recidiviz/recidiviz-data/29037
-
https://app.zenhub.com/workspaces/all-pods-6324d1d66be1c200119ab65a/issues/gh/recidiviz/recidiviz-data/29039

I've tested this on a sandbox (`_hsalas`). It works and fixes the issue.
## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7e4883edb0e2758ccab36440cd4517d360187e44