[US_OR] Correct time-served calculations for clients w/ commuted sentences (Recidiviz/recidiviz-data#28406)
## Description of the change

Corrected the `served_half_sentence` criterion for OR earned discharge
to account for cases in which a client's supervision sentence is
associated with a commuted incarceration sentence.

In these cases, we have to look back to the incarceration sentence to
determine the length of their post-prison sentence, as the sentence
length of the supervision sentence may not be accurate for the purpose
of EDIS eligibility determination. The revised criterion query uses the
original post-prison sentence length (in the incarceration sentence
data), rather than `max_sentence_length_days_calculated` for the
supervision sentence, when the associated incarceration sentence was
commuted.

As a result of this change, the number of clients currently eligible for
EDIS (according to
`task_eligibility_spans_us_or.earned_discharge_materialized`) drops
slightly, from ~7740 to ~7710.

## Related issues

Closes Recidiviz/recidiviz-data#27888

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cc7b9f905c97c8bde139dfdf6be999463a31dc3c