[US_AZ] Fix home plan preprocessed view (Recidiviz/recidiviz-data#29109)
## Description of the change
There were a few issues with the view:
1. I left the old `VIEW_BUILDER` variable in a new preprocessed view
from the one that I copied for the template. This updates the view
builder name so that the results will actually load to BQ.
2. The `base` CTE referenced `APPROVAL_DATE` instead of `UPDATE_DATE`.
That field actually represents the date that the most recent action was
taken (i.e. home plan was approved, denied, or started), not necessarily
the date it was approved, so `UPDATE_DATE` is a more accurate name for
the field.
3. A `SAFE_OFFSET` is needed in place of just an `OFFSET` in the
`status_joined_to_IP` CTE.
4. `CCL` was missing from the list of plan statuses that indicate a home
plan is in progress.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 89ef78e01749de355a141e5507ee27a407f733e5