[US_TN] Raw data fix for custody level mismatches (Recidiviz/recidiviz-data#28839)
## Description of the change

There's a small number of mismatches in the custody level we have
ingested vs. what the raw data shows. Some of these are showing up on
the front end for the TN AR/CLD tool. This PR creates
`us_tn_classification_raw` from `Classification` table in TN in analyst
views as a temporary view and unions them into `custody_level_sessions`
and `custody_level_raw_text_sessions`.

Validation notes: The person listed in the linked ticket no longer
appears as eligible for CLD. There are a few other changes between main
and sandbox that seem largely related to data lag (most of them have a
LastUpdateDate of 4/16. There are another 6 folks that are no longer
eligible for annual reclass because they appear to be in Max custody in
the raw data (but not in ingested data)

```
WITH sandbox AS (
  SELECT distinct person_external_id, 'sandbox' AS dataset,all_eligible_opportunities
  FROM `recidiviz-123.dsharm20240417_workflows_views.resident_record_materialized`,
  UNNEST(all_eligible_opportunities) AS all_eligible_opportunities
), main AS (
  SELECT distinct person_external_id, 'main' AS dataset,all_eligible_opportunities
  FROM `recidiviz-123.dsharm20240417_main_workflows_views.resident_record_materialized`,
  UNNEST(all_eligible_opportunities) AS all_eligible_opportunities
)

select *
from main
full outer join sandbox
LEFT JOIN (
  SELECT LastUpdateDate, external_id
  FROM `recidiviz-123.us_tn_raw_data_up_to_date_views.Classification_latest`
  INNER JOIN `recidiviz-123.normalized_state.state_person_external_id` pei
    ON OffenderID = pei.external_id
    AND pei.state_code = "US_TN"
  WHERE ClassificationDecision = 'A'
  QUALIFY ROW_NUMBER() OVER(PARTITION BY OffenderID
                                  ORDER BY CAST(ClassificationDate AS DATETIME) DESC) = 1
)
  ON person_external_id = external_id
using(person_external_id)
ORDER BY LastUpdateDate
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-dashboardsRecidiviz/recidiviz-data#5178

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4669ec40761b8b139a892ea0721d7790328d3e4b