[Workflows][Admin Panel] Stub opportunity configuration page (Recidiviz/recidiviz-data#29111)
## Description of the change

In the admin panel frontend, creates a Workflows store, API, and schemas
for handling opportunities and configurations. Uses the store to power a
very basic WorkflowsOpportunitiesView: my next step will be to actually
build the table for that view.

Also adds the workflows db infrastructure for running case_triage and
the admin panel backend locally. This doesn't completely work right yet,
because it loads the complete set of fixtures into every state db. I'll
see if I can adapt my js configs export script to generate per-state
CSVs we can use for fixtures.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 19f0c02b96d2fb16dc688af016b05efb70b90a64