[US_AZ] More fixes to home plan preprocessing (Recidiviz/recidiviz-data#29168)
## Description of the change

Some additional features and fixes here:
- joins to `compartment_sessions` so that you can ID the home plan
status of a particular incarceration session
- follows AZ's instructions for finding the date a home plan was
approved by referencing the `AZ_DOC_HOME_PLAN_APPROVAL` table _if and
only if_ the plan has a status that indicates it was approved.
otherwise, pulls the date of the status decision from the
`AZ_DOC_HOME_PLAN_DETAIL` page where the `STATUS_ID` comes from.
- filters to include only active rows in `AZ_DOC_HOME_PLAN`. this
doesn't exclude many rows but will be good for reliability.

Notes on these results:
- There are 2,863 rows where one person still has multiple approval
statuses with the same `update_date`. 99.5% of the time, the
`update_date` is the date of the system migration on 2019-11-30, when
all the system-generated `CREATE_DTM` values got overwritten with the
date of the migration.
- To work around this, I would suggest letting a `plan_status` of `HOME
PLAN APPROVED` trump all the others, and assume that the approval was
the most recent status wherever there is more than one on the same date.
- The other 0.5% is 12 rows, and my best guess as to what's happening
there is that there are multiple potential home plans in review for the
same person at the same time, and those independent reviews have
different results. I will ask the ADCRR data team on monday if this is
possible, or what else might be happening there.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b17b2c57d7bff7f0d1221408536b22e1f4580201