Default to enforced TLS in Sendgrid emails (Recidiviz/recidiviz-data#28932)
## Description of the change

This PR updates email delivery logic to default to the max security
settings when sending emails with PII, which is currently just Insights
emails to PA and ID.

I tested sending emails to the state users with this curl request. When
using the different settings, I updated the settings for a subuser by
using the relevant API key and making a PATCH request to the
enforced_tls API.
```
curl --request POST \
--url https://api.sendgrid.com/v3/mail/send \
--header 'Authorization: Bearer <>' \
--header 'Content-Type: application/json' \
--data '{"personalizations":[{"to":[{"email":"<recepient>"}],"subject":"Recidiviz Test Email"}],"content": [{"type": "text/plain", "value": "hello!"}],"from":{"email":"reports@recidiviz.org","name":"Recidiviz Alerts"},"reply_to":{"email":"alexa@recidiviz.org","name":"Recidiviz Support"}}'
```

When testing with state users (via a CURL request to the send API), PA
verified that they received the email and the email status was delivered
on Sendgrid when TLS settings were:
```
{
    "require_tls":true,
    "require_valid_cert":true,
    "version":1.2
}
```

Idaho user confirmed that they received the email, however, the status
on Sendgrid remained on 'Processing' with the above settings. So I
updated the settings to the below and the user confirmed they received
and Sendgrid reflected it as delivered.
```
{
    "require_tls":true,
    "require_valid_cert":false,
    "version":1.2
}
```

_Additional Context_
TLS cannot be "disabled" with Sendgrid because it opportunistically
attempts to use TLS by default. If the recipient doesn't support TLS,
then an unencrypted connection is used. However, you can enforce that
recipients support TLS with the 'Enforced TLS' feature
([docs](https://docs.sendgrid.com/ui/sending-email/tls)). To enable
sending emails with different settings, we are leveraging subusers on
Sendgrid that have different settings. Then, you can create different
API keys for each subuser.

The `RecidivizTLSAndCert` subuser has the first set of settings enforced
and you can verify this with a CURL request (see [investigation
ticket](https://github.com/Recidiviz/recidiviz-dashboards/issues/4599)
for more details) using the API key found with secret id
`sendgrid_enforced_tls_and_cert_api_key`. The `RecidivizTLSNoCert`
subuser has the second set of settings configured and the API keys can
be found with the `sendgrid_enforced_tls_api_key`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#28593

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7a3a5752e3d5d6ce4a726f56b42328ae9ae43a65