[Insights] Maintain both supervisor external id singular and plural (Recidiviz/recidiviz-data#29177)
## Description of the change

In order to prevent breaking changes, maintain both the supervisor
external id as a standalone value and as an array. This should be
deployed to each project before merging in Recidiviz/recidiviz-data#29089 so that the array
exists in the DB when the singular value is deprecated and logic is
updated to use the array. Loaded into sandbox with prefix `alexa_28250`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part one of Recidiviz/recidiviz-data#28250

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6ce9a2a4d7b7a99b802b76f7dcaa2e7cdfd8d6e8