[Insights][Analytics] Adds impact and usage metric views (Recidiviz/recidiviz-data#29035)
## Description of the change

Adds 3 views to the `outliers_views` dataset which include usage and
impact metrics to fulfill the following requirements:

  How many outlier officers were there?
How many supervisors had access? (snapshot at the end of the month is
fine)
  How many supervisors had outlier officers?
Number of supervisors with outlier officers that viewed at least 1
outlier officer this month
Number supervisors with outlier officers logged into dashboard.recidiviz
at all this month
  Number supervisors logged into dashboard.recidiviz at all this month

For each of those numbers, also generate a list of the individuals (i.e.
list of outlier officers, list of supervisors with access, etc.)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#28670

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: be6cc7985bc8d269d07a468790330c0aece23ce1