[Dataflow] Restructure how we build parameters for sandbox pipeline runs (Recidiviz/recidiviz-data#29095)
## Description of the change

A pretty serious overhaul of the PipelineParameters class and subclasses
to accomplish the following:
1) Instead of defining parameter arguments for specific input / output
datasets, users of the PipelineParameters instead take in an
`output_sandbox_dataset` and `input_dataset_overrides_json` which can be
applied to modify input / output data sources.
2) Much stricter enforcement that the pipelines write out to an output
sandbox if the pipelines indicate any sort of non-standard behavior
3) Exposes input and output dataset overrides objects that can be used
by the pipelines, e.g. the input overrides can be applied to an
arbitrary query we want to run (this is the real motivation of this
change - so I can run reference view queries with sandbox inputs).
4) Automatically applies the sandbox prefix / `-test` suffix to job
names at instantiation time now that the parameter class can detect that
it is a sandbox pipeline.

Sorry this is a big change! It was hard to not to change a bunch of
things at once. Happy to walk through this in person.

Ran the following mini-script to test:
```
python -m recidiviz.tools.ingest.development.run_sandbox_ingest_pipeline \
        --project recidiviz-staging \
        --state_code US_CA \
        --output_sandbox_prefix ageiduschek

# ...wait for pipeline to finish...

python -m recidiviz.tools.calculator.create_or_update_dataflow_sandbox \
        --project_id recidiviz-staging \
        --sandbox_dataset_prefix ageiduschek \
        --state_code US_CA \
        --datasets_to_create normalization

python -m recidiviz.tools.calculator.run_sandbox_calculation_pipeline \
        --pipeline comprehensive_normalization \
        --type normalization \
        --project recidiviz-staging \
        --job_name us-ca-normalization \
        --output_sandbox_prefix ageiduschek \
        --input_dataset_overrides_json '{"us_ca_state_primary": "ageiduschek_us_ca_state_primary"}' \
        --state_code US_CA
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#22528

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 23e126c2301691c5fdd66fbcdc5f43d6d2439efc